Couple more changes
When you click on a stop, the next view gets the name, number,
lat/long, and array of routes that go through it are sent to the
ViewController so that the information can be used to populate the view
with the correct data. Working on plotting points of stops on the map
right now.