Last update for awhile
App now centers its map on the Clinton & Iowa stop and will drop points
on stops when they are clicked on. Later we have to implement
power-saving features when determining location, but that's not
important right now. A time consuming thing will be to put in all the
right information for the stop list since there are about 97 stops, but
we'll get it done. Our next focus should be implementing the routes tab.