Clicking on a route
Clicking on the route will now show the stops that are along that
route. Unfortunately, because we want to sort the list by proximity, we
have run into an issue. The table only populates when you click on the
route, click on another tab, go back to the route tab, go back to the
route list, and then click on the route again. I'm trying to figure out
why it's doing this at the moment. Next up: plotting the stops on the
map. All that's left is to populate add the rest of the routes,
populate populate the "StopClickedOn" screen, and implement adding
favorites. This is surely not an insurmountable task for ~3 weeks left.
I'll do a little more work on this tomorrow/later tonight.