Long Live Global Variables!
I don't know why this has evaded me for so long but I made the
routesList and stopsList globally accessible variables and voila - all
the routes and and bus stops only have to load once and can be used by
other view controllers. It makes also improes performance and usability
of the app.